Use custom tabs/external browser for Uphold authentication. (#160)
Use custom tabs/external browser for Uphold authentication.  Uphold 
Integration login was failing in some scenarios which required the user
to open a link to continue the authentication process. The integration 
was using a WebView for login and opening the link on the WebView wasn't 
possible.  

To fix this, Chrome Custom Tabs was implemented to handle Uphold Login / 
Buy Dash screens, an external browser will be used as fallback when a 
browser which doesn't support Chrome Custom Tabs is available.

Add Logging to UpholdClient